Extract `File` struct from `Target` struct.
The fields `abs_path`, `owner`, `group` and `mode` in the Target struct
are all optional, and logically belong to the file the target refers to.
Since these all belong together, extract them out into a separate struct
and add an optional `file` field, renaming the existing `file` field to
`name`.

Remove the `exists()` method on Target and rework the for loop in
`main()`.

Add `error_changing_target()` method to the `Monitor` struct.