Change `Target` from a struct to an enum.
The Target enum has two variants: `Found` and `Missing`.  The Found
variant contains a `File`, the Missing variant contains only a `PathBuf`
with the name from the Options files field.  Removed now unused methods
`long_name()` and `short_name()`.

Rewrote the for loop in `main()` to switch actions based on the Target
variant.

Added a `name` field to the `File` struct, and a lifetime to go with the
name.  Partially implemented the `change_properties()` method, which now
returns a `Changes` struct on success.

Defined the Changes struct to hold options for user, group and mode.

In the mode module, defined type aliases `ModeT` and `ModeMask`.
Replaced uses of `u32` with the appropriate type alias.

In the `Monitor` struct, renamed methods:
    changed_target() --> file_was_changed()
    error_changing_target() --> file_change_failed()
    missing_target() --> target_is_missing()
    property_changes() --> change_list()
    target_name() --> file_name()

Renamed lifetimes `'n` to `'o` since both lifetimes derive from the
underlying Options struct.